fix: smooth loading screen transition
