+++ client/src/hooks/helpers/useHexPosition.tsx
@@ -15,9 +15,7 @@ export enum HexType {
 }
 
 export const useHexPosition = () => {
-  const setIsLoadingScreenEnabled = useUIStore((state) => state.setIsLoadingScreenEnabled);
   const hexData = useUIStore((state) => state.hexData);
-  const moveCameraToRealmView = useUIStore((state) => state.moveCameraToRealmView);
 
   const {
     account,
@@ -73,11 +71,6 @@ export const useHexPosition = () => {
       return hexData?.find((hex) => hex.col === neighborHex.col && hex.row === neighborHex.row);
     });
   }, [hexData, neighborHexes]);
-
-  useEffect(() => {
-    moveCameraToRealmView();
-    setIsLoadingScreenEnabled(false);
-  }, []);
 
   return {
     mainHex,
