+++ client/src/ui/modules/scenes/HexceptionViewScene.tsx
@@ -7,6 +7,7 @@ import { useControls } from "leva";
 import * as THREE from "three";
 import { HexType, useHexPosition } from "@/hooks/helpers/useHexPosition";
 import { ExistingBuildings } from "@/ui/components/construction/ExistingBuildings";
+import useUIStore from "@/hooks/store/useUIStore";
 
 const positions = {
   main: getUIPositionFromColRow(0, 0, true),
@@ -21,6 +22,9 @@ const positions = {
 export const HexceptionViewScene = () => {
   const { mainHex, neighborHexesInsideView, hexType } = useHexPosition();
 
+  const moveCameraToRealmView = useUIStore((state) => state.moveCameraToRealmView);
+  const setIsLoadingScreenEnabled = useUIStore((state) => state.setIsLoadingScreenEnabled);
+
   const texture = useTexture({
     map: "/textures/paper/paper-color.jpg",
     displacementMap: "/textures/paper/paper-height.jpg",
@@ -30,6 +34,13 @@ export const HexceptionViewScene = () => {
 
   // only realm can build
   const canConstruct = hexType === HexType.REALM;
+
+  useEffect(() => {
+    moveCameraToRealmView();
+    setTimeout(() => {
+      setIsLoadingScreenEnabled(false);
+    }, 300);
+  }, []);
 
   return (
     <>
